+++ src/1_Page/Team/ui/AwardList/ui/HoverTrophy/index.tsx
@@ -1,6 +1,6 @@
-import { useState } from "react";
 import { createPortal } from "react-dom";
 import { TeamAwardProps } from "./type";
+import useHover from "./model/useHover";
 
 const Trophy = ({
   trophyData,
@@ -9,28 +9,13 @@ const Trophy = ({
   trophyData: TeamAwardProps;
   index: number;
 }) => {
-  const [isHovered, setIsHovered] = useState(false);
-  const [hoverPosition, setHoverPosition] = useState({ x: 0, y: 0 });
-
-  const handleMouseEnter = (event: React.MouseEvent) => {
-    setIsHovered(true);
-    const rect = event.currentTarget.getBoundingClientRect();
-
-    const screenCenter = window.innerWidth / 2; // 화면 중앙
-    const offset = 60; // 추가 여백
-
-    setHoverPosition({
-      x:
-        rect.left +
-        window.scrollX +
-        rect.width / 2 +
-        (rect.left < screenCenter ? offset : -offset), // 좌우 방향 조정
-      y: rect.top + window.scrollY + rect.height, // 약간 아래로 조정
-    });
-  };
-  const handleMouseLeave = () => {
-    setIsHovered(false);
-  };
+  const [
+    isHovered,
+    hoverPosition,
+    handleHover,
+    handleMouseEnter,
+    handleMouseLeave,
+  ] = useHover();
 
   return (
     <div>
@@ -62,7 +47,7 @@ const Trophy = ({
               transform: `translate(-50%, -50%)`,
               opacity: isHovered ? 1 : 0,
             }}
-            onMouseEnter={() => setIsHovered(true)}
+            onMouseEnter={handleHover}
             onMouseLeave={handleMouseLeave}>
             <img
               src={trophyData.championship_list_throphy_img}
